Now possible to request a token without expiration or with custom expiration
Crude rate limiting for token endpoint

Disable API rate limit during tests

Typo in spec

Replaced ClientVersionRestriction and rate limiting with Rack::Attack

Tweaked Rack::Attack config, implemented simple MemoryStore, increased short keys to 128 bits.

Master auth renewal

- External OAuth2 providers can be used as authentication provider
- Internal OAuth2 provider now supports refresh token and authorization
  code grant flow
- API endpoint /authenticate is used for performing authentication
- API endpoint /v1/auth no longer available, there is no more username
  and password login
- API endpoint /v1/auth_provider is used to configure external oauth2
  provider settings
- Only encryption hashes are stored for local access and refresh tokens
- Token validation is done in RACK middleware instead of a helper module
- There's a RACK middleware that restricts access to Kontena CLI clients
  prior to version X (currently set to 0.15.0 in server.rb). The error
  message contains a hint about how to upgrade and the old CLI will show
  it when trying to perform requests.

Authentication flow:

Upon deployment a local user "admin" is created. The administrator gets
an access token with a code, generated by the CLI before deployment.

This code is then used by the CLI to perform initial authentication.

The usage of "admin" account does not require that an oauth2 provider is
configured, it's a local only account.

To configure an auth provider you can use the CLI's "kontena master
auth-provider config" command.

Authenticating a regular user:

A) Auth provider not configured:
  1) Client requests /authenticate?redirect_uri=http://localhost:xxx/cb
  2) Master responds with HTTP 500 - Authentication provider not
  configured

B) Auth provider configured
  1) Client requests
     /authenticate?redirect_uri=http://localhost:xxx/cb
     or if the user has an invitation code:
     /authenticate?redirect_uri=..&invite_code=<invite_code>
  2) Master responds with a HTTP 302 redirect to auth provider
  authorization URL.
  3) Client opens the link in a browser
  4) Once the user permits the authorization, the auth provider
  redirects back to master's /cb with an authorization code
  5) Master exchanges this code to an access token from AP's token
  endpoint.
  6) Master requests user info from AP's user info endpoint. Using the
  access token.
  7) Local user is determined by matching the external_id field.
  For first time users an invitation code is required.
  8) Master creates a local access token with an authorization code
  9) Master responds with a redirect to
  http://localhost:xxx/cb?code=<authorization_code>
  10) Client exchanges the authorization code to an access token using
  master's /oauth2/token

The Oauth2 consumer used to authenticate from external providers should
be configurable to work with almost any provider that follows the Oauth2
specification at some level. It requires that you have created an Oauth2
application on the provider and have the client_id and client_secret.

The external access tokens are currently not stored on master as it
serves no purpose because the token is not used after the authentication.
In the future it might be used to sync user data or for example to send
notifications. At that point it would be a good idea to store them in
some encrypted format or in kontena vault.

The server's oauth2 application client secret should also be stored in
such way.

Created a rake task to reset local admin

Now possible to request a token without expiration or with custom expiration

Crude rate limiting for token endpoint

Disable API rate limit during tests

Typo in spec

Replaced ClientVersionRestriction and rate limiting with Rack::Attack

Master auth renewal

- External OAuth2 providers can be used as authentication provider
- Internal OAuth2 provider now supports refresh token and authorization
  code grant flow
- API endpoint /authenticate is used for performing authentication
- API endpoint /v1/auth no longer available, there is no more username
  and password login
- API endpoint /v1/auth_provider is used to configure external oauth2
  provider settings
- Only encryption hashes are stored for local access and refresh tokens
- Token validation is done in RACK middleware instead of a helper module
- There's a RACK middleware that restricts access to Kontena CLI clients
  prior to version X (currently set to 0.15.0 in server.rb). The error
  message contains a hint about how to upgrade and the old CLI will show
  it when trying to perform requests.

Authentication flow:

Upon deployment a local user "admin" is created. The administrator gets
an access token with a code, generated by the CLI before deployment.

This code is then used by the CLI to perform initial authentication.

The usage of "admin" account does not require that an oauth2 provider is
configured, it's a local only account.

To configure an auth provider you can use the CLI's "kontena master
auth-provider config" command.

Authenticating a regular user:

A) Auth provider not configured:
  1) Client requests /authenticate?redirect_uri=http://localhost:xxx/cb
  2) Master responds with HTTP 500 - Authentication provider not
  configured

B) Auth provider configured
  1) Client requests
     /authenticate?redirect_uri=http://localhost:xxx/cb
     or if the user has an invitation code:
     /authenticate?redirect_uri=..&invite_code=<invite_code>
  2) Master responds with a HTTP 302 redirect to auth provider
  authorization URL.
  3) Client opens the link in a browser
  4) Once the user permits the authorization, the auth provider
  redirects back to master's /cb with an authorization code
  5) Master exchanges this code to an access token from AP's token
  endpoint.
  6) Master requests user info from AP's user info endpoint. Using the
  access token.
  7) Local user is determined by matching the external_id field.
  For first time users an invitation code is required.
  8) Master creates a local access token with an authorization code
  9) Master responds with a redirect to
  http://localhost:xxx/cb?code=<authorization_code>
  10) Client exchanges the authorization code to an access token using
  master's /oauth2/token

The Oauth2 consumer used to authenticate from external providers should
be configurable to work with almost any provider that follows the Oauth2
specification at some level. It requires that you have created an Oauth2
application on the provider and have the client_id and client_secret.

The external access tokens are currently not stored on master as it
serves no purpose because the token is not used after the authentication.
In the future it might be used to sync user data or for example to send
notifications. At that point it would be a good idea to store them in
some encrypted format or in kontena vault.

The server's oauth2 application client secret should also be stored in
such way.

Removed external access token expectations from oauth2 spec

Added the option to display the authorization code in browser for deployments without a local browser

Fixes to handling path exclusion in token auth

Created a rake task to reset local admin

Now possible to request a token without expiration or with custom expiration

Crude rate limiting for token endpoint

Disable API rate limit during tests

Typo in spec