Service instance deploy state, errors (#2034)
If creating the Docker container for a service instance failed during deployment, previous versions of Kontena would only log the errors on the agent, and they would not be reported back to the server or shown in the CLI. With #1607, the agent `/service_pods/create` errors would be reported back to the server, but the `GridServiceDeployer` would not return them to the CLI (#1907). With #1873, the agent  `ServicePodWorker` no longer reported the Docker API errors back to the server, and would crash instead (#2032).

This PR provides end-to-end per-instance error handling for service deployments, with the agent reporting  `ServiceInstance.error` states back to the server, the server tracks the `GridServiceInstanceDeploy.error` state, the API includes `/v1/services/:grid/:stack/:service/deploys/:deploy` `{"instance_deploys": {"error": ...}}` states, and the CLI can render those (WIP).

The `ServiceInstance` now also has a new `error `field. This is updated independently of the `state` field: a new `deploy_rev` with a `desired_state: stopped` might still result in `state: running` with `error: failed to stop...`. Then later, the agent might independently re-apply and update that `deploy_rev` to `state: stopped` `error: null`.

## Agent

* Refactor `ServicePodWorker#apply` to handle the `ensure_desired_state` success and error cases, protocol state and actor lifecycle
* Simplify `@prev_state` to just compare on the complete state update in `sync_state_to_master`

   The agent will also sync state to the master if the `error` state changes, like on a later `update` cycle.

## Server
* Add a new `error` field to the `ServiceInstance` model
* Add a new `GridServiceInstanceDeploy` model embedded in the `GridServiceDeploy`, with per-instance state and errors
* Change `GridServiceInstanceDeployer#ensure_service_instance` to update and wait for the `deploy_rev`, also when stopping an old instance on a different node

    This ensures that the instance was actually stopped. This also now warns if it was unable to stop it.
* Log warnings if unable to stop existing service instance

* Refactor the `GridServiceInstanceDeployer` around the new `GridServiceInstanceDeploy` model

## API
* Add a new `instance_deploys` field to the `/v1/services/:grid/:stack/:service/deploys/:deploy` API

## CLI
Rough hack for `kontena service deploy` to report the per-instance errors.

TODO support for `kontena stack deploy`